chore: Only timestamp rendered frames
