+++ src/Uno.UI.Dispatching/Native/NativeDispatcher.wasm.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.ComponentModel;
+using System.Diagnostics;
 using System.Linq;
 using System.Runtime.InteropServices.JavaScript;
 
@@ -9,6 +10,8 @@ namespace Uno.UI.Dispatching
 {
 	internal sealed partial class NativeDispatcher
 	{
+		private long _lastDispatchRendering;
+
 #pragma warning disable IDE0051 // Remove unused private members
 		[JSExport]
 		private static void DispatcherCallback()
@@ -75,6 +78,29 @@ namespace Uno.UI.Dispatching
 			else
 			{
 				DispatchOverride(NativeDispatcher.DispatchItems, priority);
+			}
+		}
+
+		/// <summary>
+		/// Synchronous dispatching to the dispatcher is required for Wasm.
+		/// This must be called *only* when originating from the `requestAnimationFrame` callback
+		/// </summary>
+		internal void SynchronousDispatchRendering()
+		{
+			if (Stopwatch.GetElapsedTime(_lastDispatchRendering) < DispatchingFeatureConfiguration.DispatcherQueue.FrameDuration)
+			{
+				if (this.Log().IsTraceEnabled())
+				{
+					this.Log().Trace($"Skipping frame for configured {DispatchingFeatureConfiguration.DispatcherQueue.FrameDuration}");
+				}
+				return;
+			}
+
+			_lastDispatchRendering = Stopwatch.GetTimestamp();
+
+			if (IsRendering)
+			{
+				DispatchItems();
 			}
 		}
